#617: separate timer decorator into own file; clean up writer code
